refactor(accounts,user): break accounts↔user domain cycle
Move the static isDefaultAccount and getExtendedAccounts helpers from AccountService to UserService. They read only the user's default-account fields, so they are user-domain logic; relocating them removes the lone account-service.ts → user import and breaks the accounts↔user cycle, leaving a clean one-way user → accounts dependency. Pure relocation, no behavior change. Call sites repointed: claim-cashu-token-service.ts, account-hooks.ts.

Part of the wallet-SDK no-cache migration (cycle-break before move).
