+++ apps/web-wallet/app/features/accounts/account-service.ts
@@ -6,8 +6,7 @@ import {
 import type { QueryClient } from '@tanstack/react-query';
 import type { DistributedOmit } from 'type-fest';
 import { allMintKeysetsQueryOptions } from '~/features/shared/cashu';
-import type { User } from '../user/user';
-import type { Account, CashuAccount, ExtendedAccount } from './account';
+import type { CashuAccount } from './account';
 import {
   type AccountRepository,
   useAccountRepository,
@@ -18,35 +17,6 @@ export class AccountService {
     private readonly accountRepository: AccountRepository,
     private readonly queryClient: QueryClient,
   ) {}
-
-  /**
-   * Returns true if the account is the user's default account for the respective currency.
-   */
-  static isDefaultAccount(user: User, account: Account) {
-    if (account.currency === 'BTC') {
-      return user.defaultBtcAccountId === account.id;
-    }
-    if (account.currency === 'USD') {
-      return user.defaultUsdAccountId === account.id;
-    }
-    return false;
-  }
-
-  /**
-   * Returns the accounts with the isDefault flag set to true if the account is the user's
-   * default account for the respective currency. Sorts the default account to the top.
-   */
-  static getExtendedAccounts(
-    user: User,
-    accounts: Account[],
-  ): ExtendedAccount[] {
-    return accounts
-      .map((account) => ({
-        ...account,
-        isDefault: AccountService.isDefaultAccount(user, account),
-      }))
-      .sort((_, b) => (b.isDefault ? 1 : -1)); // Sort the default account to the top;
-  }
 
   async addCashuAccount({
     userId,
